pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-todo"
description = "Todo Plugin for Alliance Auth (GitHub Version)"
readme = "README.md"
keywords = [
    "allianceauth",
    "todo",
]
license = { file = "LICENSE" }
requires-python = ">=3.10,<3.14"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.3.1,<5",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.9.0",
    "mypy>=1.8.0",
    "django-stubs[compatible-mypy]>=5.0.0",
]

[tool.hatch.build]
include = [
    "/todo",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "todo/__init__.py"

[tool.black]
line-length = 88
target-version = ["py312"]
exclude = '''
/(
    \.venv
  | venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
  "venv",
  ".venv",
  "build",
  "dist",
  "todo/migrations",
]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]
ignore = [
  "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["todo"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
mypy_path = "."
strict = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
files = ["todo"]
exclude = [
  "venv",
  ".venv",
  "build",
  "dist",
  "todo/migrations",
  "todo/tests",
]

[tool.django-stubs]
django_settings_module = "testauth.settings_mypy"

[[tool.mypy.overrides]]
module = [
  "testauth.*",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
  "celery",
  "celery.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "todo.auth_hooks",
]
ignore_errors = true
